step1 - 하위 클래스 파일 분리
